+++ examples/feature_preview/lib/knob-preview.dart
@@ -199,3 +199,33 @@ Widget buildNullableListKnob(BuildContext context) {
     ),
   );
 }
+
+@UseCase(type: KnobPreview, name: 'Object Segmented Knob')
+Widget buildObjectSegmentedKnob(BuildContext context) {
+  return OnlineStatusBadge(
+    status: context.knobs.object.segmented<OnlineStatusType>(
+      label: 'status',
+      labelBuilder: (value) => value.name,
+      options: [
+        OnlineStatusType.online,
+        OnlineStatusType.offline,
+        OnlineStatusType.busy,
+      ],
+    ),
+  );
+}
+
+@UseCase(type: KnobPreview, name: 'Object Segmented Nullable Knob')
+Widget buildObjectNullableSegmentedKnob(BuildContext context) {
+  return OnlineStatusBadge(
+    status: context.knobs.objectOrNull.segmented<OnlineStatusType>(
+      label: 'status',
+      labelBuilder: (value) => value.name,
+      options: [
+        OnlineStatusType.online,
+        OnlineStatusType.offline,
+        OnlineStatusType.busy,
+      ],
+    ),
+  );
+}